chore(Framework): Switch to .NET Framework
.NET Core must be installed on the target machine which is not as propable as
there is a .NET Framework installation.
Also Core does not mean that this app
can be called within unix systems -> WPF and Forms are still Windows
only.
Further Fody Costura is not supported in Core 3.x because of the
arising feature of a self extracting app. But this feature is not that helfpul
at the moment.